pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantization-extend"
version = "0.1.0"
description = "JPEG quantization fingerprint extraction and matching (forensic-friendly)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "manfullwel"}]
dependencies = [
  "pillow>=10.1",
  "numpy>=1.24",
  "tqdm>=4.66",
]

[project.scripts]
qext = "quantization_extend.cli:main"
